refactor: move content aggregation into feature slices
