+++ src/features/home/components/HomePage.tsx
@@ -1,9 +1,9 @@
 import BlogSection from "@/features/home/components/BlogSection";
 import HeroSection from "@/features/home/components/HeroSection";
 import PortfolioSection from "@/features/home/components/PortfolioSection";
+import { getSortedPostsData } from "@/features/blog/lib/posts";
 import SelectedWorkSection from "@/features/work/components/SelectedWorkSection";
-import { getSortedPostsData } from "@/lib/blog";
-import { getWorkSummaries } from "@/lib/work";
+import { getWorkSummaries } from "@/features/work/lib/summaries";
 
 import type { BlogSummary } from "@/features/blog/types";
